backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
pydantic-settings==2.5.0
sqlalchemy==2.0.35
alembic==1.13.2
psycopg2-binary==2.9.9
openai==1.50.0
langgraph==0.2.28
langchain-openai==0.2.1
celery==5.4.0
redis==5.1.0
python-dotenv==1.0.1
python-multipart==0.0.12
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
